Apply unified button text style
