+++ assets/js/dashboard/components/popover.tsx
@@ -29,7 +29,8 @@ const toggleButton = {
     shadow:
       'bg-white dark:bg-gray-800 shadow text-gray-800 dark:text-gray-200 hover:bg-gray-200 dark:hover:bg-gray-900',
     ghost:
-      'text-gray-500 hover:text-gray-800 hover:bg-gray-200 dark:hover:text-gray-200 dark:hover:bg-gray-900'
+      'text-gray-500 hover:text-gray-800 hover:bg-gray-200 dark:hover:text-gray-200 dark:hover:bg-gray-900',
+    truncatedText: 'truncate block font-medium'
   }
 }
